% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01_clean_enroll.R
\name{clean_enroll}
\alias{clean_enroll}
\title{clean_enroll()}
\usage{
clean_enroll(enroll_data, program_type, start_date = NULL,
  end_date = Sys.Date())
}
\arguments{
\item{enroll_data}{dataframe: a dataframe containing enrollment data."[Admin] raw_enrollment_report"}

\item{program_type}{dataframe: a dataframe containing information about each program intervention type. "[Admin] raw_program_type_report"}

\item{start_date}{character: All program start dates in enroll_data must be posterior or equal to this date. Date must be specified as "mm/dd/yyyy"}

\item{end_date}{character: All program end dates in enroll_data must be anterior or equal to this date. Date must be specified as "mm/dd/yyyy"}
}
\value{
dataframe
}
\description{
This function takes two dataframes exported from the following ETO Results reports: 1)"[Admin] raw_enrollment_report", and 2) "[Admin] raw_program_type_report" as inputs. It does some specific cleaning, merge them, and returns one dataframe ready for further analysis.
}
\examples{
enroll <- laycUtils::load_txt('./my_data_folder/enrollment.txt')
enroll <- laycUtils::clean_data(enroll)
ptype <- laycUtils::load_txt('./my_data_folder/program_type.txt')
ptype <- laycUtils::clean_data(ptype)
clean_enroll(enroll_data = enroll, program_type = ptype)
}

