% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_enroll.R
\name{get_enroll}
\alias{get_enroll}
\title{get_enroll()}
\usage{
get_enroll(enroll_data, eto_programs = NULL)
}
\arguments{
\item{enroll_data}{dataframe: a dataframe containing enrollment data."[Admin] raw_enrollment_report"}

\item{eto_programs}{character vector:  a vector of character containing the name of ETO programs to keep for analysis.}
}
\value{
numeric
}
\description{
This function returns the number of ubduplicated participants who enrolled in a specific program or group of programs. It takes one dataframe as input. The data comes from the following ETO Results report: "[Admin] raw_enrollment_report".
}
\examples{
enroll <- laycUtils::load_txt('./my_data_folder/enrollment.txt')
enroll <- laycUtils::format_data(enroll)

get_enroll(enroll_data = enroll)
}

