% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_enrollment.R
\name{get_enrollment}
\alias{get_enrollment}
\title{get_enrollment()}
\usage{
get_enrollment(enroll_data, program_type, start_date = "01/01/2008",
  end_date = Sys.Date())
}
\arguments{
\item{enroll_data}{dataframe: a dataframe containing enrollment data.[INSERT REPORT NAME HERE]}

\item{program_type}{dataframe: a dataframe containing information about each program intervention type. [INSERT REPORT NAME HERE]}

\item{start_date}{character: All program start dates in enroll_data must be posterior or equal to this date. Date must be specified as "mm/dd/yyyy"}

\item{end_date}{character: All program end dates in enroll_data must be anterior or equal to this date. Date must be specified as "mm/dd/yyyy"}
}
\value{
list
}
\description{
This function takes two dataframes 1)[INSERT REPORT NAME HERE], and 2) [INSERT REPORT NAME HERE] as inputs, and returns summary tables and charts about participants enrollment
}
\examples{
enroll <- laycUtils::load_txt('./my_data_folder/enrollment.txt')
ptype <- laycUtils::load_txt('./my_data_folder/program_type.txt')
get_enrollment(enroll_data = enroll, program_type = ptype)
}

