% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/03_get_histogram.R
\name{get_histogram}
\alias{get_histogram}
\title{get_histogram()}
\usage{
get_histogram(df)
}
\arguments{
\item{df}{dataframe: a dataframe returned by the get_summary() function}
}
\value{
dataframe
}
\description{
Create a multiple histograms representing enrollment length by programt type (workforce, education, etc.)
}
\examples{
enroll <- laycUtils::load_txt('./my_data_folder/enrollment.txt')
ptype <- laycUtils::load_txt('./my_data_folder/program_type.txt')
df <- clean_df(enroll_data = enroll, program_type = ptype)
summary <- get_summary(df)
get_histogram(summary)
}

