% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02_get_summary.R
\name{get_summary}
\alias{get_summary}
\title{get_summary()}
\usage{
get_summary(df)
}
\arguments{
\item{df}{dataframe: a dataframe returned by the clean_df() function}
}
\value{
dataframe
}
\description{
Compute summary statistics about enrollment, by program type (education, workforce, etc.)
}
\examples{
enroll <- laycUtils::load_txt('./my_data_folder/enrollment.txt')
ptype <- laycUtils::load_txt('./my_data_folder/program_type.txt')
df <- clean_df(enroll_data = enroll, program_type = ptype)
get_summary(df)
}

