% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03_plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{plot_summary()}
\usage{
plot_summary(df)
}
\arguments{
\item{df}{dataframe: a dataframe returned by the get_summary() function}
}
\value{
dataframe
}
\description{
Plot summary statistics about enrollment
}
\examples{
enroll <- laycUtils::load_txt('./my_data_folder/enrollment.txt')
ptype <- laycUtils::load_txt('./my_data_folder/program_type.txt')
df <- clean_df(enroll_data = enroll, program_type = ptype)
summary <- get_summary(df)
plot_summary(summary)
}

