% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/01_clean_df.R
\name{clean_df}
\alias{clean_df}
\title{clean_df()}
\usage{
clean_df(enroll_data, program_type, start_date = NULL,
  end_date = Sys.Date())
}
\arguments{
\item{enroll_data}{dataframe: a dataframe containing enrollment data.[INSERT REPORT NAME HERE]}

\item{program_type}{dataframe: a dataframe containing information about each program intervention type. [INSERT REPORT NAME HERE]}

\item{start_date}{character: All program start dates in enroll_data must be posterior or equal to this date. Date must be specified as "mm/dd/yyyy"}

\item{end_date}{character: All program end dates in enroll_data must be anterior or equal to this date. Date must be specified as "mm/dd/yyyy"}
}
\value{
dataframe
}
\description{
This function takes two dataframes 1)[INSERT REPORT NAME HERE], and 2) [INSERT REPORT NAME HERE] as inputs, does some cleaning, merge them, and returns one dataframe ready for further analysis.
}
\examples{
enroll <- laycUtils::load_txt('./my_data_folder/enrollment.txt')
ptype <- laycUtils::load_txt('./my_data_folder/program_type.txt')
clean_df(enroll_data = enroll, program_type = ptype)
}

