% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_enrollment.R
\name{get_summary}
\alias{get_summary}
\title{get_summary()}
\usage{
get_summary(df)
}
\arguments{
\item{df}{dataframe: a dataframe returned by the clean_df() function}
}
\value{
dataframe
}
\description{
Helper function to compute summary statistics about enrollment, by program type (education, workforce, etc.)
}

